requirements.txt
mysql-connector-python
tkinter
matplotlib
pandas